--- a/dask/tests/test_utils.py
+++ b/dask/tests/test_utils.py
@@ -417,17 +417,18 @@ def test_ndeepmap():
 def test_ensure_dict():
     d = {"x": 1}
     assert ensure_dict(d) is d
-    hlg = HighLevelGraph.from_collections("x", d)
-    assert type(ensure_dict(hlg)) is dict
-    assert ensure_dict(hlg) == d
 
     class mydict(dict):
         pass
 
-    md = mydict()
-    md["x"] = 1
-    assert type(ensure_dict(md)) is dict
-    assert ensure_dict(md) == d
+    d2 = ensure_dict(d, copy=True)
+    d3 = ensure_dict(HighLevelGraph.from_collections("x", d))
+    d4 = ensure_dict(mydict(d))
+
+    for di in (d2, d3, d4):
+        assert type(di) is dict
+        assert di is not d
+        assert di == d
 
 
 def test_itemgetter():